requirements.txt
# Core dependencies with pinned versions for compatibility
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.1

# Supabase - specific versions known to work together
supabase==2.4.0  # Changed from 2.2.0 to 1.0.1
httpx==0.24.1  # Exact version that works with supabase 1.0.1
gotrue==2.0.0  # Must match with supabase 1.0.1
storage3==0.5.3
realtime==1.0.0

# Authentication
python-dotenv==1.0.1
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.9

# Testing
pytest==8.0.2

# UI
streamlit==1.32.2
pillow==10.2.0
watchdog==3.0.0